refactor: extract xtdb transport
